chore: move verify_password and get_password_hash to utils file
