+++ eirb-connect/app/utils.py
@@ -4,7 +4,23 @@ This module contains helper functions
 
 import base64
 from app.conf import mongodb
+from passlib.context import CryptContext
 
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+# Helper password functions
+def verify_password(plain_password, hashed_password):
+    """
+    Helper function to check if a password matches a hashed password
+    """
+    return pwd_context.verify(plain_password, hashed_password)
+
+
+def get_password_hash(password):
+    """
+    Helper function to generate a hashed password
+    """
+    return pwd_context.hash(password)
 
 def encrypt_service(service_url: str) -> str | None:
     """
